requirements.txt
ffmpeg-python
faster-whisper